Conflicts:
	jenkins/plugin/src/main/java/com/hp/octane/plugins/jenkins/configuration/JobConfigurationProxy.java
	mqm-client/mqm-rest-client/src/main/java/com/hp/mqm/client/model/TaxonomyType.java

+++ mqm-client/mqm-rest-client/src/main/java/com/hp/mqm/client/model/Release.java
@@ -4,15 +4,15 @@ package com.hp.mqm.client.model;
 
 final public class Release {
 
-    final private int id;
+    final private Long id;
     final private String name;
 
-    public Release(int id, String name) {
+    public Release(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
